Fix recursive directory push for DeviceAsync class
